backend-daemon/requirements.txt
funasr
numpy
scipy
pyzmq

